+++ packages/cli-tools/bin/streamr-storage-node-list-streams.ts
@@ -10,7 +10,7 @@ createClientCommand((async (client: StreamrClient, storageNodeAddress: string) =
         console.info(EasyTable.print(streams.map((stream) => {
             return {
                 id: stream.id,
-                partitions: stream.partitions
+                partitions: stream.getMetadata().partitions
             }
         })))
     }