+++ packages/client/src/registry/StreamRegistry.ts
@@ -7,11 +7,10 @@ import { scoped, Lifecycle, inject, delay } from 'tsyringe'
 import { EthereumConfig, getAllStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
 import { until } from '../utils/promises'
 import { ConfigInjectionToken, TimeoutsConfig } from '../Config'
-import { Stream, StreamProperties } from '../Stream'
-import { ErrorCode, NotFoundError } from '../HttpUtil'
+import { Stream, StreamMetadata } from '../Stream'
+import { NotFoundError } from '../HttpUtil'
 import { StreamID, StreamIDUtils } from 'streamr-client-protocol'
 import { StreamIDBuilder } from '../StreamIDBuilder'
-import { omit } from 'lodash'
 import { SynchronizedGraphQLClient } from '../utils/SynchronizedGraphQLClient'
 import { searchStreams as _searchStreams, SearchStreamsPermissionFilter } from './searchStreams'
 import { filter, map } from '../utils/GeneratorUtils'
@@ -95,8 +94,8 @@ export class StreamRegistry {
     }
 
     private parseStream(id: StreamID, metadata: string): Stream {
-        const props: StreamProperties = Stream.parsePropertiesFromMetadata(metadata)
-        return this.streamFactory.createStream({ ...props, id })
+        const props = Stream.parseMetadata(metadata)
+        return this.streamFactory.createStream(id, props)
     }
 
     private async connectToContract(): Promise<void> {
@@ -111,25 +110,8 @@ export class StreamRegistry {
         }
     }
 
-    async getOrCreateStream(props: { id: string, partitions?: number }): Promise<Stream> {
-        try {
-            return await this.getStream(props.id)
-        } catch (err: any) {
-            if (err.errorCode === ErrorCode.NOT_FOUND) {
-                return this.createStream(props)
-            }
-            throw err
-        }
-    }
-
-    async createStream(propsOrStreamIdOrPath: StreamProperties | string): Promise<Stream> {
-        const props = typeof propsOrStreamIdOrPath === 'object' ? propsOrStreamIdOrPath : { id: propsOrStreamIdOrPath }
-        props.partitions ??= 1
-
+    async createStream(streamId: StreamID, metadata: StreamMetadata): Promise<Stream> {
         const ethersOverrides = getStreamRegistryOverrides(this.ethereumConfig)
-
-        const streamId = await this.streamIdBuilder.toStreamID(props.id)
-        const metadata = StreamRegistry.formMetadata(props)
 
         const domainAndPath = StreamIDUtils.getDomainAndPath(streamId)
         if (domainAndPath === undefined) {
@@ -146,7 +128,7 @@ export class StreamRegistry {
                 know what the actual error was. (Most likely it has nothing to do with timeout
                 -> we don't use the error from until(), but throw an explicit error instead.)
             */
-            await waitForTx(this.streamRegistryContract!.createStreamWithENS(domain, path, metadata, ethersOverrides))
+            await waitForTx(this.streamRegistryContract!.createStreamWithENS(domain, path, JSON.stringify(metadata), ethersOverrides))
             try {
                 await until(
                     async () => this.streamExistsOnChain(streamId),
@@ -158,12 +140,9 @@ export class StreamRegistry {
             }
         } else {
             await this.ensureStreamIdInNamespaceOfAuthenticatedUser(domain, streamId)
-            await waitForTx(this.streamRegistryContract!.createStream(path, metadata, ethersOverrides))
+            await waitForTx(this.streamRegistryContract!.createStream(path, JSON.stringify(metadata), ethersOverrides))
         }
-        return this.streamFactory.createStream({
-            ...props,
-            id: streamId
-        })
+        return this.streamFactory.createStream(streamId, metadata)
     }
 
     private async ensureStreamIdInNamespaceOfAuthenticatedUser(address: EthereumAddress, streamId: StreamID): Promise<void> {
@@ -173,19 +152,17 @@ export class StreamRegistry {
         }
     }
 
-    async updateStream(props: StreamProperties): Promise<Stream> {
-        const streamId = await this.streamIdBuilder.toStreamID(props.id)
+    // TODO maybe we should require metadata to be StreamMetadata instead of Partial<StreamMetadata>
+    // Most likely the contract doesn't make any merging (like we do in Stream#update)?
+    async updateStream(streamId: StreamID, metadata: Partial<StreamMetadata>): Promise<Stream> {
         await this.connectToContract()
         const ethersOverrides = getStreamRegistryOverrides(this.ethereumConfig)
         await waitForTx(this.streamRegistryContract!.updateStreamMetadata(
             streamId,
-            StreamRegistry.formMetadata(props),
+            JSON.stringify(metadata),
             ethersOverrides
         ))
-        return this.streamFactory.createStream({
-            ...props,
-            id: streamId
-        })
+        return this.streamFactory.createStream(streamId, metadata)
     }
 
     async deleteStream(streamIdOrPath: string): Promise<void> {
@@ -274,10 +251,6 @@ export class StreamRegistry {
             }
         }`
         return JSON.stringify({ query })
-    }
-
-    private static formMetadata(props: StreamProperties): string {
-        return JSON.stringify(omit(props, 'id'))
     }
 
     private static buildGetStreamWithPermissionsQuery(streamId: StreamID): string {